requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
gunicorn==21.2.0
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
openpyxl==3.1.5
reportlab==4.2.5
Pillow==10.4.0
